Get rid of 'evernote-oauth' and install 'omniauth'
Removing the 'evernote-oauth' gem because I couldn't figure out how to
get it to work. Decided to go with OmniAuth to handle OAuth because I
had a good experience with it in the past and figured it would be
simpler to set up.

Also installing the 'omniauth-evernote' provider to enable
authentication with Evernote.

Comment out the route I had created to `auth/evernote#index` for
'evernote-oauth'.