Add 'kipcole9/evernote' gem
In searching for an easier way to interface with the Evernote SDK and to
get all notes for a given notebook, I came across a wrapper gem for the
Evernote Ruby SDK: https://github.com/cgs/evernote.

The trouble is, this gem was last updated on June 15, 2012, and is
likely out of date. Checked the GitHub Network tab to see if there were
any forks a little farther along. Turns out kipcole9 has one that has
some nice updates, and is the most recent of all the forks.

Installing kipcole9's version to help me get some note data.