NotesController: Display notes and note contents
For all notebooks, display all notes inside a notebook and the contents
and tags of each note.

Rendering the note contents with `html_safe` to get them to display
quickly in a user-readable form because the contents are formatted in
ENML, Evernote's XML markup language.

Tried collecting the notes in the controller but ultimately decided to
output them directly in the view.